feat(lib): Acommodate contract method in the base of the contract
